Completamento Refactoring per package "manager"
