+++ Android/app/src/main/java/com/fsc/cicerone/ItineraryCreation.java
@@ -20,6 +20,7 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.android.volley.Request;
 import com.android.volley.toolbox.StringRequest;
+import com.fsc.cicerone.manager.ItineraryManager;
 
 import org.json.JSONException;
 import org.json.JSONObject;
