+++ Android/app/src/main/java/com/fsc/cicerone/LoginActivity.java
@@ -15,6 +15,7 @@ import android.widget.EditText;
 import android.widget.RelativeLayout;
 import android.widget.Toast;
 
+import com.fsc.cicerone.manager.AccountManager;
 import com.fsc.cicerone.model.UserType;
 
 import org.json.JSONException;
